Fix TODO's in tests
There is no migration test currently with embedded ddb
But we do test Lease refresher with embedded ddb that when
createGSI and waitForGSIToBeActive is invoked it creates-but-not-wait
and waits-for-gsi-to-be-active respectively. So its good enough
to ensure those APIs are invoked from migration state machine.
So just added a few more verifications to existing tests.